Remove 'Standby' instances when deleting ASG
When removing an Auto Scaling Group, should one or more instances be in
'Standby' mode, the ec2_asg will hang indefinitely waiting for all
instances to be gone (group.instances == []).

This is normally accomplished by setting max_size, min_size, and
desired_capacity all to 0, which causes the ASG to terminate all
instances.

Instances with lifecycle_state of 'Standby' are not terminated by the
above action.  Using the AWS Console, these instances are deleted.  To
mirror this behavior, we iterate over the instances member, and
specifically terminate those with life_cycle_state of 'Standby'.